backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.0.76
numpy==1.24.3
pytesseract==0.3.10
Pillow==10.0.0
Werkzeug==2.3.7 